Fixed memory readout
Basically this commit fixes the memory readout for the daemon. Before it was only reading the "free" free memory, which is often used up completely by the OS and turned into "available memory", which is a pre processed memory which is ready to be allocated to a new process, while pre processed, it is still free. This fix will also make the bot read the available memory and count it as free memory too.